bootstrap like menu and IRAM_ATTR

+++ lib/WiFiSettings/WiFiSettings.cpp
@@ -93,7 +93,7 @@ bool WiFiSettings::getStorageSizeIsAvailable()
     return this->storageSizeIsAvailable;
 }
 
-bool WiFiSettings::eraseSettings(uint16 startAddress, uint size) {
+bool WiFiSettings::eraseSettings(uint16_t startAddress, uint16_t size) {
   if (this->getStorageSizeIsAvailable())
   {
     uint16_t lastAddress = startAddress + size;
